Fix the direct variable indexing inside (sub)frame

+++ libavcodec/amrwbdata.h
@@ -54,17 +54,19 @@ enum Mode {
     NO_DATA                                ///< no transmission
 };
 
+/* All decoded parameters in these structs must be 2 bytes long
+ * because of the direct indexing at the frame parsing */
 typedef struct {
     uint16_t adap;                         ///< adaptive codebook index
-    uint8_t ltp;                           ///< ltp-filtering flag
-    uint8_t vq_gain;                       ///< VQ adaptive and innovative gains
-    uint8_t energy;                        ///< high-band energy
+    uint16_t ltp;                          ///< ltp-filtering flag
+    uint16_t vq_gain;                      ///< VQ adaptive and innovative gains
+    uint16_t energy;                       ///< high-band energy
     uint16_t pul_ih[4];                    ///< MSBs part of codebook index (high modes only)
     uint16_t pul_il[4];                    ///< LSBs part of codebook index
 } AMRWBSubFrame;
 
 typedef struct {
-    uint8_t vad;                           ///< voice activity detection flag
+    uint16_t vad;                          ///< voice activity detection flag
     uint16_t isp_id[7];                    ///< index of ISP subvectors
     AMRWBSubFrame subframe[4];             ///< data for subframes
 } AMRWBFrame;
